refactor: move oci artifact_definitions to crate::oci
The agent-type/package OCI artifact vocabulary is not package-specific;
move it to the crate::oci layer so non-package consumers (the upcoming
agent-type downloader) don't depend on `package`.
