+++ agent-control/src/oci/artifact_definitions.rs
