+++ agent-control/src/package/oci.rs
@@ -1,3 +1,2 @@
-pub mod artifact_definitions;
 pub mod downloader;
 pub mod package_manager;
